video/verilator: use mach_defines.h GFX_* constants
This is a little dodgy: using an include file for the completed system
(ipl/gloss/mach_defines.h) to test a low level component (soc/video). This is
to set us up for writing code in a form that will also run on the completed
system, and thus posibly serve as examples.