video/verilator: Use GFX_REG() style access
Use GFX_REG() and GFX* arrays to access video memory, in the same way that
sdk.h works. This will allow the creation of example code which runs in
verilator and is close to being able to be cut and paste to app code.

Also updates existing code to use GFX_REG() and GFX* arrays as appropriate.