playing around with multiple voices
changing the voice type on the fly seems to prevent routing/synthesis
this is a bummer.  but OTOH, using a ton of voices seems to be no
problem
this suggests taking voice type as a parameter, assigning them
statically.  probably makes sense anyway.

path of least resistance: config a bunch of voices, let the users go
nuts