qpimem_dma_rdr.v: fix display with certain pitch lengths
Addresses the case where framebuffer pitch lengths that were not a multiple of 16
pixels would cause the linerenderer to omit filling up the buffer every 30
lines, causing 4 lines to repeat and then overflowing the frame. The result was
a single elongated frame being shown over 2 frames, with much flicker.

Was this a bug? Maybe, maybe not. However, having allowing multiples of 4 seems
better than only allowing multiples of 16.

The cause of the behavior appears to be that fifo_wptr and fifo_rptr were
aliases of 9 bits each of the qpi_addr and out_addr bit vectors. By having
pitches that differed by 4n+1, 2, or 3, these could match at times when it was
perhaps not intended that they should, causing dma_ready to go false and
halting further processing. I'm sure there's more to it, but this is as far as
I got in debugging.

This patch turns fifo_wptr and fifo_rptr into regs, and always starts DMAs with
fifo_wptr = fifo_rptr = 0. This seems to sidestep whatever problems were.