tool-settings.mk: Make work with WSL+Linux tools
When using Windows Subsystem for Linux, search for the RiscV toolchain compiled
either for Windows or for Linux. The Linux version is preferred, and if not
found falls back to the Windows version.

Under WSL. the linux tools compile helloworld in 1.5s, while the Windows tools
take 4.6s to do the same.