Decreased conditions for amix_out_cur to amix_out assignment.
The alphamixer_out <= alphamixer_out_cur can be done outside of the
state machine as it is an output dependent on the vid_wen bit. It very
slightly improves the timing for clk48m from 49.36 MHz to 49.74 MHz.

I will be adding another combinatorial step in that output path as part
of the workshop. In that case this change make a huge timing difference
fail vs pass. Having this minor change will make the workshop bit less
convoluted and easier to follow.