Working around weird make mkdir issue on Mac OS X
For some reason, $(1) was always returning the build directory and $(2) would
always be the app directory inside the build_template recipe. This caused the
mkdir to try and make ./build even when it needed to make ./build/subdir.

This is fixed, or at least made to be more consistent by using the directory of
the output file rather than the templated variable. There may be a different
fix but I couldn't figure it out. The foreach loop does seem to pass in the
correct path to the template function.