New app: hackaday supercon 2019 animator
This uses color cycling techniques to simulate some of the awesome
displays we've seen at the conference.  The source material for the
image was the tote bag, then converted to 1-bit and with each element
colored to a different color.  This was saved as a 75-color PNG
and run through optipng to get it to use a compact palette.